+++ data/src/main/java/com/example/data/datasource/db/dao/TaskDao.kt
@@ -17,18 +17,18 @@ interface TaskDao {
     suspend fun updateTask(task: Task)
 
     @Query("SELECT * FROM task_table WHERE task_id = :id")
-    suspend fun getTaskWithId(id: Int): Task?
+    suspend fun getTaskWithId(id: Long): Task?
 
     @Query("DELETE FROM task_table WHERE task_id = :id")
-    suspend fun removeTaskWithId(id: Int)
+    suspend fun removeTaskWithId(id: Long)
 
     @Query("UPDATE task_table SET task_pinned = 1 WHERE task_id = :id")
-    suspend fun pinTask(id: Int)
+    suspend fun pinTask(id: Long)
 
     @Query("UPDATE task_table SET task_pinned = 0 WHERE task_id = :id")
-    suspend fun unPinTask(id: Int)
+    suspend fun unPinTask(id: Long)
 
-    suspend fun taskExist(taskId: Int): Boolean = getTaskWithId(taskId) != null
+    suspend fun taskExist(taskId: Long): Boolean = getTaskWithId(taskId) != null
 
     @Query("SELECT * FROM task_table")
     suspend fun getAllTasks(): List<Task>
